requirements.txt
streamlit==1.32.0
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
protobuf==3.20.3 